Implemented Batch Execution from input file of commands.
User can supply an optional command-line argument which
is a file name. Each line in the file can be a command
or a comment, beginning with a #, or a blank line.
The database will execute the commands during initialization.
Also wrote a basic internal test for predicate handling in
test_query_evaluator. Haven't tested more sophisticated
predicates yet.